Added an internal compiler type dtLeaderToken which is used to instantiate leader iterators for the purposes of resolving forall loops.  This is then used to mark the iterator and its wrappers as requiring the early inlining to handle the parallel and locality constructs used within it.  Previously to this change, if the leader iterator was wrapped, the iterator itself would not be inlined. Depending on where the leader argument is located, this could be a problem.  In particular, leader iterator methods would not work.
As a side effect, this allows a user-invocation of a leader iterator
to be treated differently than a compiler-invocation.  This is
important if we don't want a programmer to invoke a faux-leader
iterator in order to use cobegin, on, etc. in it.  This would cause
major problems if the faux-leader iterator were zippered.

Added a new working test case of a leader iterator that is wrapped.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14567 3a8e244f-b0f2-452b-bcba-4c88e055c3ca