Add the counted range operator
The stride of the result is the same as the stride of the operand.
When applied to a bounded range with positive stride, count elements are taken
starting from the low bound. On a bounded range with negative stride, count
elements are taken starting from the high bound.

On a boundedLow range, the low bound is fixed and the high bound is set based
on the count and absolute value of the stride.

On a boundedHigh range, the high bound is fixed and the low bound is set based
on the count and absolute value of the stride.

Updated some hpcc/ra and hpcc/fft tests to (unfortunately) explicitly cast the
low bound of some ranges to be the same type as the count since the range
bounds are no longer parameters when they get to the count operator. When
possible, these should be changed back.

Updated my counted and strided range test to use the standard operators
instead of defining its own.

Updated the spec to make this an operator instead of part of a literal range.
Changed BoundedType back to BoundedRangeType after it was somehow reverted in
revision 14402.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14569 3a8e244f-b0f2-452b-bcba-4c88e055c3ca