Improved destructors (Paired with Samuel)
Added FnSymbol* Type::destructor that points to a destructor for any
given type after function resolution.

Added code in function resolution to resolve a destructor for every
class, record, and union.

Added code in buildDefaultFunctions to build a default destructor for
every class, record, and union that doesn't have one.

Moved the call to CHPL_FREE out of the destructor so that it is called
at the place where 'delete' is invoked, but after the destructor is
called.

Moved code in normalize that chains destructors to the end of function
resolution.

Improved destructors in the standard modules to make them more
generally applicable.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14571 3a8e244f-b0f2-452b-bcba-4c88e055c3ca