Reworked domain code to request distribution to make domains
Note: One expected regression: test/arrays/marybeth/CMO_array.chpl

Yesterday, I learned that we didn't request that a distribution
create a new domain -- rather, we asked the distribution what
type of class it would make if we asked it to make a domain and
then asked that domain class to start making copies of itself.
Among other things, this caused us to allocate a lot of extra
distribution classes (Sam's remaining 4 bytes of memory in FFT)
and made it very difficult to write interesting user-defined
distributions.

Reworked the code that parses a domain type to deal more in
terms of distribution class instances and tuck them into
a _build_domain_type() call.  Then, in cleanup.cpp, find
these calls and rewrite them based on context.  If they are
part of a domain const/variable's declared type, they are
moved to the initializer field (folding the user's initializer
into the call) and translated into a call to a new call:
_dist_build_domain().  If they are in any other context, this
same call is created, but then surrounded by a PRIMITIVE_TYPEOF
to convert it into a type.  This is nice in that it doesn't
require the distribution or ChapelArray.chpl to use any type
functions for this purpose, and avoids duplication of logic
for how to create a new domain.  In fact, I believe that this
will permit some other things to be removed from the 
ChapelArray.chpl module (buildEmptyDomain?, wrapDomain?, ...?).

Made an abstract Distribution class that defines all of the
possible newDomaain() methods for arithmetic, associative,
enumerated, opaque domains such that if a particular Distribution
class doesn't override them, the user will get a useful
(runtime) error message -- worth looking at whether this could
somehow be made into a compile-time error message.  (I believe
that Steve has also bristled at the concepts of using an
abstract Distribution class before, so we might need to look
at that as well).

Renamed our old "SingleLocaleDistribution" class to "DefaultDist"
and made it support newDomain() calls for all of the possible
domain types.  Created a single instance of this called
"DefaultDistribution" which all domains will use by default
if they don't provide their own distribution class.

Made user constructors for all of the SingleLocale*Domain
classes to avoid the requirement to use repetitive name-based
constructor calls when building these class types.  Removed
initialize() methods for these classes as well.  This makes
the "rank" fields of the opaque/associative domains less
annoying to deal with (and maybe they can go away if/when
wrapDomain does?)

Removed some stray "buildOpenIntervalUpper()" calls that had
never been removed.

Added a compilerError for trying to add indices to an
arithmetic domain.

Made ChapelLocale use SingleLocaleArithmetic since it does,
to ensure proper compilation order.


Next steps:

- get CMO_array.chpl working again

- can we remove buildEmptyDomain (or at least the this=nil
  version of it?  Or at the very least make it a method on
  a distribution?)

- I don't think the sparse stuff is quite in line with the
  above strategy, though it is working.  Worth another look
  once the block distribution is a bit further along.

- Look into changing the abstract Distribution methods into
  compilerErrors (or equivalent) rather than halt()s.  Look
  into removing the abstract Distribution class?

- Move Distribution and DefaultDist and DefaultDistribution
  into their own file (and come up with better names?)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14574 3a8e244f-b0f2-452b-bcba-4c88e055c3ca