Minor distribution cleanup
Through my own experimentation and Steve's insight, made the
_dist_build_domain calls be generic w.r.t. their distribution
argument rather than making it a "Distribution".  This has
two main effects:  

(1) allows compilerError()s to be used in the 'abstract' 
    methods in the Distribution class, permitting a nice 
    compile-time error message to be generated when a
    domain is created for a distribution that doesn't support 
    it (added tests of these messages);

(2) reduces the number of methods that need to be supported
    by an abstract distribution at compile time since the
    link between a distribution class and the domain record
    is done on an instantiation-only basis rather than
    through inheritance.

Removed the warning code that we used to generate when someone
tried to use distributions -- this wasn't working with my new
checkin, and is not completely trivial to re-enable; since
it's going away real soon now, I'd rather just live without it
for now and deal with the single GASNet regression.

Removed a field from the DefaultDist that I added in a moment
of paranoia yesterday and forgot to remove.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14583 3a8e244f-b0f2-452b-bcba-4c88e055c3ca