Checking in my next steps toward making BradsBlock1DPar into a distribution
This is still in .notest mode, but gets us much closer to using an
"official" distribution to specify the implementation of
stream-block1d's distribution, domain, and arrays.  In particular, for
this check-in, I grabbed the current copy of BradsBlock1DPar.chpl and
massaged it into the distribution framework as best I could.  The
point at which things fell apart was the buildEmptyDomain() call which
only really makes sense for non-distributed distributions (or at
least, those for which the domain's global class is sufficient for
creating a new domain -- generally it's not for distributed domains).
Steve is currently working on introducing a runtime domain type that
will address this issue (at which point, we'll build empty domains by
requesting them from the distribution).

In particular, when asked to create an emptyDomain, I have the option
of either setting its distribution to nil (causing a core dump in the
compiler currently due to an attempt to access the distribution's
targetLocDom field in an array member); or setting it to a completely
bogus distribution (which gets us past that point, but causes a seg
fault later because the distribution doesn't describe anything useful).

The specific steps I took to wrestle BradsBlock1DPar.chpl into an
official distribution (which might be useful to David down the road
when he's doing the same for his cyclic and block-cyclic work) were
as follows.

* made my distribution class a subclass of Distribution

* made the newDomain() call match the expected interface:
  def newDomain(param rank, type idxType, param stridable: bool)

* made versions of newDomain for ranks/index types that I'm not
  prepared to handle yet (should've done stridable too -- I drop
  that on the floor and have a TODO suggesting I fix that someday).

* made my global domain class a subclass of BaseDenseArithmeticDomain

* added a param rank field, required by the _domain record

* moved code to set up local domain classes into a helper routine, as
  it is now called both from the constructor and setIndices()

* renamed newArray() to be buildArray() (even though I like the name
  newArray better -- will petition for this later)

* added routines required by _domain record:
  - buildEmptyDomain() -- this should go away after the runtime domain
                          type is implemented; and it is the reason this
                          code doesn't currently work -- it's not possible
                          to create a reasonable distributed domain without
                          a distribution class instance
  - setIndices() -- I did two versions, one that takes a domain, the other
                    takes the tuple of ranges as in the SingleLocaleArith...
  - getIndices()

* made my global array class a subclass of BaseArray.

* renamed the array's element type from elemType to eltType as required
  by the _array record wrapper.


(These may also be of interest to see where my default namings
differed from those that we're currently using):



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14585 3a8e244f-b0f2-452b-bcba-4c88e055c3ca