Protecting against seg faults
Checking in a few checks/assertions to head off a seg fault that can
occur when trying to generate a field of a nil instance variable.
This happened for me when declaring a class with members along the
lines of:

      class D {
        var x: C;
        var y: [C.foo] real;
      }

If x is nil, I got a seg fault when the compiler was generating
the class constructor.  Now it gives an internal error.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14588 3a8e244f-b0f2-452b-bcba-4c88e055c3ca