Fix the regressions caused by r14592
In that revision, I added conditionals to only link with GASNet if
MAKE_LAUNCHER == 0. The problem was that when compiling codes with launcher
type none, this variable is unset.

Instead, when compiling the launcher binary, set MAKE_LAUNCHER = 1 and in
the locations where it was previously compared to 0, check that it is not 1.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14595 3a8e244f-b0f2-452b-bcba-4c88e055c3ca