Made the compiler support for runtime array types into a much more general mechanism to support arbitrary runtime types via a pragma 'has runtime type' and two type functions that must be defined for these runtime types.  Used this general mechanism to implement runtime domain types that contain their distribution and runtime sparse domain types that contain their parent domain.
There is an expected regression of release/examples/varargs which I'll
try to fix shortly.  This test passes a domain type to a function that
takes a variable length argument list of types; the runtime type seems
to be lost.

Changed the parent domain for sparse domains into an actual domain, as
opposed to a BaseDenseArithmeticDomain.  This simplifies the
implementation nicely and seemed necessary during the above process.

Removed all special-casing of sparse domains from the compiler
(normalize, scopeResolve, and function resolution) now that they have
a runtime type.

Removed special-casing of domain types from cleanup.

There is still some work to simplify arrays, domains, and
distributions as a result of these changes.  In this change, I did
change some function names to make them more clear and orthogonal.

Removed rank parameter from domain classes other than arithmetic.  For
example, removed it from SingleLocaleAssociativeDomain.

With these changes, the SingleLocaleSparseDomain 'this' instance is
never nil when buildEmptyDomain is called so I was able to remove that
check.  I plan to see if buildEmptyDomain can now be removed
altogether in a separate checkin.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14596 3a8e244f-b0f2-452b-bcba-4c88e055c3ca