Fixed a bug in the implementation of domain runtime types for opaque domains.  Opaque domains require extra handling because the idxType passed to the chpl_buildDomainRuntimeType function is not the same as the idxType of opaque domains.  This results in two distinct runtime types for the same array, and possible type mismatches in the generated code.  Fixed this issue by creating a wrapper version of chpl_buildDomainRuntimeType that looks for opaque and calls the function on _OpaqueIndex.
I added a comment about this.  The problem is that we should have only
a single type associated with opaque domains, not two.  I believe the
reason we have two is so that opaque can be used for interfacing with
C.  The problem is that opaque, as used in opaque domains, is a
distinct type, so we make the idxType for opaque domains different.  I
think this fix is fairly robust so we can continue to have two
distinct notions of opaque but this is odd and took a while to figure
out.

This checkin should fix the new GASNet regressions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14603 3a8e244f-b0f2-452b-bcba-4c88e055c3ca