Implemented initial follower for RandomStream class
I implemented a parallel-safe follower iterator for the RandomStream
class.  It's still far from perfect, but a step forward.  

In particular, the limitations are:

(a) the indices that the leader generates are used as indices into the
RandomStream, rather than ofsetting them by the current "cursor" value
of the stream and reading off the next logical k values;

(b) related, the stream is not updated after the followers are used,
so the next reads will occur from wherever they would've occurred
prior to the parallel iteration over the stream.

This permitted me to remove the future I filed earlier today and file
a second future that points out these new limitations (by using the
same random stream to fill two 10 element arrays and then comparing to
a 20 element array).

My implementation approach here was nifty, though potentially
suboptimal (though potentially not?).  Each follower iterator
invocation allocates its own local copy of the RandomStream class to
prevent races with the original class and the overhead of non-local
accesses to the global class.

Both the cursor-oriented reuse problems above and some of the
overheads of this approach (in particular, the allocation of a new
class for each follower invocation rather than reusing one copy for
multiple followers) could be solved by some sort of startup/teardown
hooks for the follower iterator invocations as Steve and I have
been discussing (also potentially of use to avoid 0-basing all
indices yielded by the leader).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14626 3a8e244f-b0f2-452b-bcba-4c88e055c3ca