Changed arithmetic array used in DefaultAssociative module into a tuple, saving about 0.5 seconds when compiling jacobi.
Renamed my new chpl_* functions in modules to chpl__*.

Reworked mangling of local variables so that temps are named T1, T2,
...  Local variables are not mangled unless necessary, and in that
case become T#_theirname where # is the next available #.  This is a
work in progress and I'll be making some more changes soon.  The idea
is to not use the compiler id except when specified by a flag, and in
that case, the compiler id will be output in a comment.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14633 3a8e244f-b0f2-452b-bcba-4c88e055c3ca