Removing Andy's index + domain future
It turns out I forgot to add the +(index, domain) case
which is what Andy asked for -- added it here.  I still
have not added the -(index, domain) case because it's
not trivial to implement and I'm not confident whether
or not it makes sense.  Opinions?  I guess we support
it for ranges, so in removing *, / from ranges, I can
reconsider it there.

Removed one test that I wrote which relied on index + domain
resulting in promotion, but it was more cutesy than anything
deep.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14638 3a8e244f-b0f2-452b-bcba-4c88e055c3ca