Removed * and / on ranges
Removed the * and / operators on ranges.  With the
introduction of the # operator, they are not nearly
as necessary, and they always resulted in certain
ambiguities/issues that nobody was terribly comfortable
with:  e.g., what does rng * 0 mean?  What does rng * -1
mean?  What does (1..n by 1) / 2 mean?  At the time of
their introduction, I felt that supporting these was more
important than not because it cleanly supported the ability
to take variable-length strided slices of arrays as needed
for FFT and the like (e.g., (0..radix) * stride + low)
but now that we have the # operator, this can be written
even more cleanly as low.. by stride #radix.

I still believe + and - are important on ranges due to
their intuitiveness, cleanness, and the fact that they
keep the operations closed on ranges without any confusing
cases.

The one remaining open question is whether int - range
should be supported.  Need to wrestle with that one still.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14640 3a8e244f-b0f2-452b-bcba-4c88e055c3ca