Removed -(x:integral, r:range)
After discussion with Steve, decided not to support
"integral - range" any longer.  It's a slightly strange
operator in that it will turn degenerate ranges into
non-degenerate and vice-versa (unless the range only
has one element).  Not very intuitive or obviously
useful.

Both versions of + (integral + range AND range + integral)
are still supported under the argument that + is a
commutative operator while - is not.

I think this is the end of my range/domain operator
changes for now.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14641 3a8e244f-b0f2-452b-bcba-4c88e055c3ca