various codegen cleanup
Added flag --gen-ids to the developer debugging category.  Using this
causes the compiler ids of symbols to be printed out in comments
before their definition point.  Removed --short-names which is now the
default for all temps named _tmp.

Added newTemp(name, type) function that returns a new VarSymbol with
FLAG_TEMP set and with the specified name and type.  Both arguments
are optional and the default is "_tmp" for name and dtUnknown for
type.  Note that the name can be left off and the type specified
(because this is overloaded).

Renamed symbols associated with _class_id.

Changed chpl__primes tuple into a function to avoid a bottleneck for
associative domains since the constant is located on Locale 0.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14643 3a8e244f-b0f2-452b-bcba-4c88e055c3ca