Made Random follower iterator take a 64-bit domain
Due to the bug I just filed, the Random follower was
having trouble following the zippered iteration in
the stream benchmark because the Block1D distribution
was yielding int(64) domains but the Random leader
was (arbitrarily) yielding an int(32) domain.

Switched Random to use int(64) domains for the time
being until the bug is fixed, futurized the
parRandom2-domblock.chpl test that I just checked
in under the same future, and added a 64-bit version
that works correctly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14651 3a8e244f-b0f2-452b-bcba-4c88e055c3ca