Update the varargs example and add futures that show the bugs it caused
The original mode of failure was not initializing a domain in a tuple because
of the runtime type. This happens with runtimeTypeInTuple2.chpl.

Recently, varargs started failing with a runtime seg fault like the test
runtimeTypeInTuple.chpl.

I tried changing the domain in the tuple to the type "object". This caused a
runtime seg fault from an attempt to access the class id field of a nil class
for dynamic dispatch. This is captured in nilDynamicDispatch.chpl.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14661 3a8e244f-b0f2-452b-bcba-4c88e055c3ca