Check in a label+continue future test
This is the test I mentioned when arguing that label statements should be
removed from the language.

The outer loop doesn't have its index variable updated after the inner loop
continues, so the result is an infinite loop with i always set to 1. I added a
count to terminate the otherwise infinite loop.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14664 3a8e244f-b0f2-452b-bcba-4c88e055c3ca