Adding -Wno-uninitialized to the gcc generated code compiler flags
Steve pointed out last night that -Winitialized (turned on by -Wall)
is conservative and therefore will warn about things that we know
are defined before they're used.  I had (very naively) been thinking
that initializing the variables to make the warnings go away would
be OK because the C compiler could optimize the redundant assignments
away but of course if it could prove that it could do this, it wouldn't
give the warning in the first place.  Steve went on to point out that
valgrind will warn about true use-before-def errors, making this
flag somewhat redundant with our valgrind runs (except that we don't
yet do valgrind runs for GASNet or --no-local compilations, but we
could/should).  So, in the interests of making the compiler simpler
and the generated code more efficient, we decided to try taking out
this flag and seeing where it gets us.  Steve may remove other
initializations that came about simply because of this flag as well.

This allowed me to remove my leadFollowOpt future, and to add -O to
my stream-block1d performance run.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14666 3a8e244f-b0f2-452b-bcba-4c88e055c3ca