Reimplementing - on domains
My previous implementation of "domain - index" simply
negated the index values and used the existing 
"translate" call to perform the subtraction.  I hadn't
anticipated at the time that this would cause problems
for uint domains because uint values don't support unary
negation.  Oops!

This check-in fixes the issue by implementing a
"reverse translate" call on domains and ranges that is 
like translate, but uses subtraction rather than
addition.  Since it wasn't clear that we wanted to
reveal this to the user as a method on ranges/domains,
named this chpl__unTranslate for this checkin.

Added a test that makes sure we can do + and - on
uint domains to lock this capability in.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14673 3a8e244f-b0f2-452b-bcba-4c88e055c3ca