Changing type of range returned by # operator
The previous version of the # operator returned a range
with the same eltType as the input range.  With this
check-in, I make the eltType of the result be the type
of the input eltType + the integer value being passed in.
Thus 0..m where m is an int(64) now generates an int(64)
range rather than an error message.  0..n where n is a
uint(64) still generates an error message (though a
different one) since int(32) can't be added to uint(64).

Added a test of these two cases, where the latter is a
future indicating that we should improve our error messages
for illegal combinations of integers to # by specializing
them in the ChapelRange module rather than having them
fall out due to lower-level function resolution.

Updated the spec to specify what the index type of the #
operator is.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14675 3a8e244f-b0f2-452b-bcba-4c88e055c3ca