Adding Block1D version of RA
This check-in contains a first working draft of a version of
RA that (a) uses the Block1D distribution and (b) uses a
follower iterator in the ra-randstream.chpl file in order to
write the main loop as a forall loop.

Had to make some changes to Block1D to do some type casting
in some of my internal math in order to make it legal with
uint index types.  Boy are uints annoying...




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14676 3a8e244f-b0f2-452b-bcba-4c88e055c3ca