Making a copy of the stream-block1d.chpl for --local performance testing
This is a copy of my stream-block1d.chpl test that uses the
--local flag to see how that affects the performance of the
code.

Perhaps I should change the performance testing system (and
the general testing system?) to test a program for *each* 
line in a .compopts/.execopts file so that multiple compile
commands could be applied to a single test without replicating
all the other files?



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14679 3a8e244f-b0f2-452b-bcba-4c88e055c3ca