Making Random follower iterator re-callable
Have each follower iterator update the global cursor when it
completes executing.  This makes my stream codes run correctly
on one locale, but for small problem sizes, not on multiple
locales because of the chance that two followers will run
consecutively, causing one of them to see the updated cursor
position rather than the original one.  I believe that the
leader/follower interface will need to change in order to fix
this.

I'm not removing my stream-block1d*.futures for this reason.
In fact, for the checked-in problem size, they almost always
fail on multiple locales.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14682 3a8e244f-b0f2-452b-bcba-4c88e055c3ca