Make calling a method with a nil this reference a runtime error
Insert a nil check on the first argument of class method calls and getcid
calls.

Remove a check for "this == nil" in the associative array module. It is
guaranteed to be false by the nil check that happens before calling the method.
None of our tests had a problem with this.

Update the Writer.write method to call the "nil".writeThis method when it
encounters a nil object instead of relying on the object's writeThis method
to check for nil.

Update a few tests to make an instance of an object before calling methods on
it and fixed a few line numbers in .good files.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14684 3a8e244f-b0f2-452b-bcba-4c88e055c3ca