Next performance improvement to Block1D
This one does a dynamic check at the top of the local array
follower to determine whether all the indices being followed
are local to "here".  If they are, a specialized version of
the follower loop is run which does all of the accesses using
the local array descriptor rather than the global array
descriptor.

The next improvements to this would be to:

1) put this loop within a local block.  However, this doesn't
   work due to the semantic future I filed yesterday in which
   local blocks within follower iterators seem to get dropped
   on the floor.

2) declare the myLocArr variable to be a "local" class instance
   since we know that it's local to "here" by definition.  This
   doesn't work because we've never implemented the ability to
   declare a class instance as "local".

My next step is going to be to implement these by hand and see
how close they get us to our other stream performance numbers.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14685 3a8e244f-b0f2-452b-bcba-4c88e055c3ca