Backing out my change to Random.chpl from yesterday
In making the Random module update the global cursor yesterday,
I knew that it resulted in a race condition for parallel iteration,
but forgot that there were non-futurized tests that would fail due
to this race.  Backing out the update to the cursor and adding a
comment as a placeholder.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14686 3a8e244f-b0f2-452b-bcba-4c88e055c3ca