Don't allow iteration over boundedNone ranges
In the range these iterator, halt if the range is boundedNone. This is similar
to the behavior when the range is boundedLow with a negative stride or
boundedHigh with a positive stride.

Add some tests that don't behave well without this change.
badRange.chpl assigns a variable the value ".. .." which caused a runtime
infinite loop.
badRange2.chpl promotes a function with an int argument for a boundedNone
range. This caused it to choose 1 as a start value.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14688 3a8e244f-b0f2-452b-bcba-4c88e055c3ca