Implemented fillRandom on reals to use parallel iteration
Got the distributed stream benchmarks working as we would like
by changing the explicit calls to the RandomStream follower iterator
into calls to fillRandom() and making the fillRandom() call itself
be parallel.  This has several advantages:

1) It's the way we wanted to write the codes anyway; the explicit
   forall versions were simply workarounds until we got the
   fillRandom call to work in parallel.

2) It avoids the various race and correctness issues we were
   wrestling with for adding leaders/followers to the RandomStream
   class and allows us to put off the proposed expansion(s) to
   the leader/follower interface that we've been discussing.

3) It allows the tests to work correctly and in parallel.

My implementation approach was to change the (public) these iterators
that I had defined for RandomStream in the past few weeks into
internal-only iterators.  I then made the things that were challenging
about supporting these iterators (like anticipating the number of
things that the sequential iterator would have to return; and updating
the shared cursor value in a sane way) get handled by the fillRandom call
itself, which is easy since it's a sequential routine.

I futurized the tests that were explicitly testing the RandomStream
follower as a feature request because I think it's something we'll
want someday.  I also made a variation on the one that tested that
the execution really was parallel using the fillRandom routine and
a config param in the Random module.  This revealed an unfortunate
bug (?) in our module reference-without-use code that I had to
work around by useing the Random module in the top-level fillRandom()
call.  I'll file a future against this workaround in my next checkin.

TODO:
- do the same for the complex fillRandom




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14697 3a8e244f-b0f2-452b-bcba-4c88e055c3ca