Added an error check to guard against returning nested iterators or loop expressions since these can capture a reference to a local variable in a function (and the function may return).  Changed the .good for
  modules/figueroa/DuplicateConfigVars/RandomNumber

to test this error.  Added a rewrite of this test to capture the loop
expression into a variable (and then return that).

It is still an open question as to how we should handle this longer
term.  We need to decide when things should be put on the heap and
when things should result in errors because things are not put on the
heap.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14700 3a8e244f-b0f2-452b-bcba-4c88e055c3ca