Reimplemented the heap allocation portion of the compiler.  This used to be done in normalize before function resolution.  It is now done in the parallel pass after function resolution.
The bulk of the new heap allocation code is in a new function called
makeHeapAllocations that analyzes the entire program in order to
identify variables that need to be put on the heap, change the types
of these variables, allocates them on the heap, and update the code so
that it is still valid given these changes.

Added code in the expansion of iterator loops that moves the index
into the loop so that if it is put on the heap there will be a
separate copy for each iteration.

Added a new flag FLAG_HEAP_ALLOCATE used to mark variables that need
to be put on the heap.  This is used in buildCoforallLoopStmt to mark
its indices as needing to go on the heap.

Added a new flag FLAG_NEW_HEAP_TYPE to mark the heap type that is
inserted by the compiler in this new heap allocation pass.  There is a
significant amount of dead code related to the old heap type.  I plan
to remove that shortly and then rename this flag to FLAG_HEAP.

This change should fix the following --no-local/gasnet regressions
which were a result of complications due to putting things on the heap
before function resolution:

  classes/deitz/tree.chpl
  classes/diten/deleteobject.chpl
  functions/deitz/iterators/test_promote_inout.chpl
  functions/deitz/iterators/tree.chpl
  memory/shannon/memtrackFlag.chpl
  modules/diten/returnClassDiffModule5.chpl
  parallel/sync/diten/tupleOfSync.chpl
  portability/boolSizes.chpl
  release/examples/iterators.chpl
  types/scalar/bradc/overloadbools2.chpl
  types/scalar/bradc/overloadbools.chpl


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14701 3a8e244f-b0f2-452b-bcba-4c88e055c3ca