Rework the PBS launcher to use interactive batch mode
As Mike Karo suggested, the launcher binary creates and runs an expect script.
The script executes qsub in interactive mode, waits for a prompt, changes
to the work directory, executes aprun and waits for another prompt.

Standard input is passed through to the _real binary using an expect command,
so input is possible.

The PBS launcher looks at the CHPL_PROJECT_STRING environment variable and
passes whatever it is set to to qsub.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14718 3a8e244f-b0f2-452b-bcba-4c88e055c3ca