Squelching info messages in Cray XT compiles
By default, the Cray XT compiler prints out an info line with
every invocation of the compiler targeting a CNL node.  This
is annoying and can be squelched by setenv'ing
XTPE_INFO_MESSAGE_OFF to true, but rather than require each
user to do this in each XT environment, why not do it in our
Makefiles?  Seems worth a shot anyway.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14724 3a8e244f-b0f2-452b-bcba-4c88e055c3ca