Removing requirement that CHPL_COMM be set before building gasnet
My last checkin assumed that the user would have CHPL_COMM set to
gasnet before building the third-party/gasnet directory.  This is
a poor assumption.  This change fixes this by always setting
CHPL_COMM to gasnet when building the gasnet directory, and by
exporting the CHPL_COMM variable to the util/ scripts when called
by the Makefile.comm script.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14732 3a8e244f-b0f2-452b-bcba-4c88e055c3ca