Implemented reference counting for domains
Replaced the _locked field for domain records with _count, which is intended to
keep track of how many arrays refer to the same domain.  The destructor
decrements the count and frees memory if the count reaches zero.  No doubt there
are currently places where references are copied without incrementing _count, so
probably in many cases, _count will reach zero even though there are still
references to the corresponding domain.  In any case, the compiler currently
does not insert calls to destructors for domains.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14733 3a8e244f-b0f2-452b-bcba-4c88e055c3ca