Changed forall to for to support --enable-parallel-iterators
Mary Beth's HPL code used a forall loop over a user-defined
iterator, prior to our invention of the leader-follower
scheme.  Rather than rewrite this iterator to be a leader/follower
(or better, a parallel standalone iterator?), I changed it to
a forall and commented that it could/should be a forall for
future reference.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14737 3a8e244f-b0f2-452b-bcba-4c88e055c3ca