Making RA tests work with --enable-parallel-iterators
In particular, I am disabling the ra-hpcc06 test (which should've
been done a long time ago) because it was meant to be a snapshot
of what we submitted in 2006 to see when it stopped working, not
a code that was meant to be kept up-to-date.  Eventually, we
should back out the changes that have already been made to this
code.

Changed the forall in ra-paramWidth to coforall to match our other
ra tests.  We may want to retire the subBlocks iterator, either
permanently or until we have the distribution interface defined
well enough to be sure that this will be a part of it.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14740 3a8e244f-b0f2-452b-bcba-4c88e055c3ca