Made it so that all leader iterators are inlined by marking the leader iterator itself (by looking for the leader tag).  We used to inline all leader iterators by marking the forall loop.  Made it so that the compiler does not lower leader iterators.  This used to happen but was disabled yesterday.  This should fix the baseline regressions.
This change is also important moving forward since there are a number
of places in the modules where we wrap leader iterators and invoke
them explicitly with the tag.  This should probably be legal (even
non-internal), but we will eventually want to improve the error
message for the case where the user tries to zipper a leader iterator
with another iterator.  This will currently break (probably
ungracefully) only if the leader iterator contain parallel constructs.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14747 3a8e244f-b0f2-452b-bcba-4c88e055c3ca